+++ sulong/CHANGELOG.md
@@ -6,6 +6,8 @@ New features:
   If not, a _mismatching target triple_ error is reported. To make this error non-fatal,
   set `--llvm.verifyBitcode=false`. To silence the message, set `--log.llvm.BitcodeVerifier.level=OFF`.
 
+* Allow accessing foreign buffers from LLVM bitcode via the interop buffers API.
+
 # Version 21.0.0
 
 Fixes: